fixed shop bug, moved some methods from game to map
