+++ DeadZone/source/Game/Game.cpp
@@ -33,7 +33,6 @@ Game::Game()
     : MAX_NUM_DEAD_BODIES(100) //daca sunt 100 de dead body-uri pe jos atunci incepem sa stergem in ordinea cronologica
 {
     WindowManager::get();
-    genUtil = GeneralUtilities::get();
 
     // TODO: mutat in main
     // Initialize ENet
@@ -179,7 +178,7 @@ void Game::loadResources()
     {
         int width = gameJSON["map"]["width"].get<int>();
         int height = gameJSON["map"]["height"].get<int>();
-        std::string file = this->generateProceduralMap(width, height);
+        std::string file = Map::get().generateProceduralMap(width, height);
 
         Map::get().readMap(file);
     }
@@ -393,130 +392,4 @@ void Game::updateRemotePlayerPosition(const std::string& clientKey, double x, do
     remotePlayers[clientKey]->setY(y);
 }
 
-void Game::putDoorsInEnclosedAreas(const int& width, const int& height, std::vector<std::vector<std::string>>& map, std::vector<std::vector<bool>>& enclosed) {
-   std::vector<std::vector<bool>> visited(height, std::vector<bool>(width, 0));
-    const int di[4] = { -1, 0, 1, 0 };
-    const int dj[4] = { 0, 1, 0, -1 };
 
-    auto inside = [&](std::pair<int, int> cell) {
-        if (cell.first < height && cell.first >= 0 && cell.second < width && cell.second >= 0)
-            return 1;
-        return 0;
-        };
-
-    std::queue<std::pair<int, int>> cellsInQueue;
-    std::vector<std::pair<int, int>> candidatesForDoor;
-    std::vector<std::pair<int, int>> visitedCells;
-    for (int i = 0; i < height; i++)
-        for (int j = 0; j < width; j++)
-            if (map[i][j][0] == '.' && !visited[i][j]) {
-                candidatesForDoor.clear();
-                visitedCells.clear();
-                int cnt = 0;
-                cellsInQueue.push({ i, j });
-                visited[i][j] = 1;
-                visitedCells.push_back({ i, j });
-                while (cellsInQueue.size() > 0) {
-                    cnt++;
-                    std::pair<int, int> cell = cellsInQueue.front();
-                    cellsInQueue.pop();
-                    for (int k = 0; k < 4; k++) {
-                        std::pair<int, int> new_cell = { cell.first + di[k], cell.second + dj[k] };
-                        if (inside(new_cell) && !visited[new_cell.first][new_cell.second]) {
-                            if (map[new_cell.first][new_cell.second][0] == '.') {
-                                cellsInQueue.push({ new_cell.first, new_cell.second });
-                                visited[new_cell.first][new_cell.second] = 1;
-                                visitedCells.push_back({ new_cell.first, new_cell.second });
-                            }
-                            else if (map[new_cell.first][new_cell.second][0] == 'M')
-                                candidatesForDoor.push_back({ new_cell.first, new_cell.second });
-                        }
-                    }
-                }
-                if (cnt < 100) {
-                    for (int k = 0; k < visitedCells.size(); k++) {
-                        std::pair<int, int> cell = visitedCells[k];
-                        enclosed[cell.first][cell.second] = 1;
-                    }
-                    for (int k = 0; k < candidatesForDoor.size(); k++) {
-                        std::pair<int, int> cell = candidatesForDoor[k];
-                        if (cell.first > 0 && cell.first < height - 1 && cell.second > 0 && cell.second < width - 1) {
-                            map[cell.first][cell.second] = "D0";
-                            break;
-                        }
-                    }
-                }
-            }
-}
-
-void Game::putShopInGoodArea(const int& width, const int& height, std::vector<std::vector<std::string>>& map, const std::vector<std::vector<bool>>& enclosed) {
-    const std::vector<std::vector<std::string>> pattern = { {".2", ".6", ".1"}, {".8", ".9", ".7"}, {".8", ".9", ".7"}, {".8", ".9", ".7"}, {".4", ".5", ".3"} };
-
-    std::vector<std::pair<int, int>> shopPrefered, shopAnyway;
-    for (int i = 0; i < height - 6; i++)
-        for (int j = 0; j < width - 3; j++) {
-            bool okSP = 1, okSA = 1;
-            for (int k = 0; k < 6; k++)
-                for (int l = 0; l < 3; l++) {
-                    if (map[i + k][j + l][0] != '.')
-                        okSP = 0, okSA = 0;
-                    if (enclosed[i + k][j + l] == 0)
-                        okSP = 0;
-                }
-            if (okSP)
-                shopPrefered.push_back({ i, j });
-            if (okSA)
-                shopAnyway.push_back({ i, j });
-        }
-    std::pair<int, int> positionForShop;
-    if (shopPrefered.size() > 0) {
-        int ind = Random::randomInt(1, static_cast<int>(shopAnyway.size())) - 1;
-        positionForShop = shopPrefered[ind];
-    }
-    else {
-        int ind = Random::randomInt(1, static_cast<int>(shopAnyway.size())) - 1;
-        positionForShop = shopAnyway[ind];
-    }
-    std::cout << positionForShop.first << ' ' << positionForShop.second << '\n';
-    for (int i = 0; i < 5; i++)
-        for (int j = 0; j < 3; j++)
-            map[i + positionForShop.first][j + positionForShop.second] = pattern[i][j];
-    map[5 + positionForShop.first][1 + positionForShop.second] = "shop0";
-}
-
-std::string Game::generateProceduralMap(const int& width, const int& height) {
-
-    std::vector<std::vector<std::string>> map(height, std::vector<std::string>(width, "."));
-
-    // Generate Perlin Noise Map
-    genUtil.generatePerlinMap(width, height, 10, 256, map);
-
-    // Update corners
-    for (int i = 0; i < height; i++)
-        map[i][0] = map[i][width - 1] = "M0";
-
-    for (int j = 0; j < width; j++)
-        map[0][j] = map[height - 1][j] = "M0";
-
-    // Put doors where areas are enclosed
-    std::vector<std::vector<bool>> enclosed(height, std::vector<bool>(width, 0));
-    putDoorsInEnclosedAreas(width, height, map, enclosed);
-
-    // Find position for shop
-    putShopInGoodArea(width, height, map, enclosed);
-
-    // Gettime sinch epoch in ms
-    long long ms_since_epoch = genUtil.getTimeSinceEpochInMs();
-
-    std::string maps_dir = "maps/";
-    std::string new_map_name = "sandbox_" + std::to_string(ms_since_epoch);
-    std::string output_dir = maps_dir + new_map_name + ".map";
-
-    // Output generated file
-    std::ofstream MAP_OUTPUT(output_dir);
-    for (int i = 0; i < height; i++, MAP_OUTPUT << "\n "[i == height])
-        for (int j = 0; j < width; j++)
-            MAP_OUTPUT << map[i][j] << ' ';
-
-    return output_dir;
-}
