+++ DeadZone/source/Game/Game.h
@@ -9,7 +9,6 @@
 #include "../Entity/Entity.h"
 #include "../Entity/DeadBody/DeadBody.h"
 #include "../Entity/RemotePlayer/RemotePlayer.h"
-#include "../GeneralUtilities/GeneralUtilities.h"
 
 class Game
 {
@@ -48,7 +47,6 @@ public:
 
 private:
 	GameStatus gameStatus = GameStatus::InMenu;
-	GeneralUtilities genUtil;
 
 public:
 	static Game& get();
@@ -70,10 +68,5 @@ public:
 
 	void spawnRemotePlayer(const std::string& clientKey);
 	void updateRemotePlayerPosition(const std::string& clientKey, double x, double y);
-
-	void putDoorsInEnclosedAreas(const int& width, const int& height, std::vector<std::vector<std::string>>& M, std::vector<std::vector<bool>>& enclosed);
-	void putShopInGoodArea(const int& width, const int& height, std::vector<std::vector<std::string>>& map, const std::vector<std::vector<bool>>& enclosed);
-
-	std::string generateProceduralMap(const int& width, const int& height);
 };
