+++ DeadZone/source/Map/Map.h
@@ -11,6 +11,8 @@
 #include "../Entity/Door/Door.h"
 #include "../Entity/Shop/Shop.h"
 
+#include "../GeneralUtilities/GeneralUtilities.h"
+
 /*
 * Codificare:
 * 
@@ -22,7 +24,7 @@
 class Map
 {
 private:
-	Map() = default;
+	Map();
 	Map(const Map& other) = delete;
 	Map& operator= (const Map& other) = delete;
 	Map(const Map&& other) = delete;
@@ -37,6 +39,8 @@ private:
 	std::vector<std::shared_ptr<Door>> doors;
 	std::vector<std::shared_ptr<Shop>> shops;
 
+	GeneralUtilities genUtil;
+
 public:
 	~Map() = default;
 	static Map& get();
@@ -49,5 +53,9 @@ public:
 	static void deleteInstance();
 
 	bool hasBeenLoaded() const { return mapLoaded; }
+
+	void putDoorsInEnclosedAreas(const int& width, const int& height, std::vector<std::vector<std::string>>& M, std::vector<std::vector<bool>>& enclosed);
+	void putShopInGoodArea(const int& width, const int& height, std::vector<std::vector<std::string>>& map, const std::vector<std::vector<bool>>& enclosed);
+	std::string generateProceduralMap(const int& width, const int& height);
 };
